feat: tab key at the last table cell creates a row (#1095)
