+++ cypress/support/index.js
@@ -16,6 +16,7 @@
 // Import commands.js using ES2015 syntax:
 import './commands';
 import 'cypress-plugin-snapshots/commands';
+require('cypress-plugin-tab');
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
