+++ package.json
@@ -69,6 +69,7 @@
     "cssnano": "4.1.11",
     "cypress": "^9.2.0",
     "cypress-plugin-snapshots": "^1.4.4",
+    "cypress-plugin-tab": "^1.0.5",
     "cz-lerna-changelog": "^2.0.2",
     "docz": "1.3.2",
     "docz-theme-default": "1.2.0",
