+++ packages/rich-text/src/plugins/Table/createTablePlugin.ts
@@ -1,26 +1,16 @@
 import { BLOCKS, CONTAINERS } from '@contentful/rich-text-types';
-import {
-  getBlockAbove,
-  getParent,
-  HotkeyPlugin,
-  KeyboardHandler,
-  getLastChildPath,
-  WithPlatePlugin,
-  getAbove,
-} from '@udecode/plate-core';
+import { getBlockAbove, getParent, getLastChildPath, WithPlatePlugin } from '@udecode/plate-core';
 import {
   createTablePlugin as createDefaultTablePlugin,
   ELEMENT_TABLE,
   ELEMENT_TD,
   ELEMENT_TH,
   ELEMENT_TR,
-  onKeyDownTable,
   withTable,
 } from '@udecode/plate-table';
 import { NodeEntry, Path, Transforms } from 'slate';
 
 import { isRootLevel } from '../../helpers/editor';
-import { insertEmptyParagraph } from '../../helpers/editor';
 import { transformLift, transformParagraphs, transformWrapIn } from '../../helpers/transformers';
 import { RichTextPlugin, CustomElement, RichTextEditor } from '../../types';
 import { addTableTrackingEvents } from './addTableTrackingEvents';
@@ -30,44 +20,13 @@ import { Row } from './components/Row';
 import { Table } from './components/Table';
 import { createEmptyTableCells, getNoOfMissingTableCellsInRow, isNotEmpty } from './helpers';
 import { insertTableFragment } from './insertTableFragment';
-
-const createTableOnKeyDown: KeyboardHandler<RichTextEditor, HotkeyPlugin> = (editor, plugin) => {
-  const defaultHandler = onKeyDownTable(editor, plugin as WithPlatePlugin);
-
-  return (event) => {
-    // This fixes `Cannot resolve a Slate point from DOM point: [object HTMLDivElement]` when typing while the cursor is before table
-    const windowSelection = window.getSelection();
-    if (windowSelection) {
-      // @ts-expect-error
-      const blockType = windowSelection.anchorNode.attributes?.['data-block-type']?.value; // this attribute comes from `plugins/Table/components/Table.tsx`
-      const isBeforeTable = blockType === BLOCKS.TABLE;
-
-      if (isBeforeTable) {
-        if (event.key === 'Enter') {
-          const above = getAbove(editor, { match: { type: BLOCKS.TABLE } });
-
-          if (!above) return;
-
-          const [, tablePath] = above;
-
-          insertEmptyParagraph(editor, { at: tablePath, select: true });
-        }
-
-        event.preventDefault();
-        event.stopPropagation();
-        return;
-      }
-    }
-
-    defaultHandler(event);
-  };
-};
+import { onKeyDownTable } from './onKeyDownTable';
 
 export const createTablePlugin = (): RichTextPlugin =>
   createDefaultTablePlugin({
     type: BLOCKS.TABLE,
     handlers: {
-      onKeyDown: createTableOnKeyDown,
+      onKeyDown: onKeyDownTable,
     },
     withOverrides: (editor, plugin) => {
       // injects important fixes from plate's original table plugin
